fix: resolve peer-programming code-review findings
Code-review sweep follow-ups for the peer-programming plugin:

- #1074: removed the duplicate dead _lib.ts (app/api copy) and made the
  surviving lib/ copy map all error codes (invalid_payload/policy_denied/
  not_found/assignment_not_found) instead of falling through to 503.
- #1075: POST /admin/single-open-cohort treats a missing `enabled` key the
  same as null (clear the setting); an empty body {} no longer returns 400.
- #1077: enforce PEER_PROGRAMMING_MAX_MESSAGE_LENGTH on POST /messages and
  POST /messages/[messageId]/replies (400 when over the limit).
- #1080: enforce PEER_PROGRAMMING_MAX_FEEDBACK_LENGTH on POST /feedback.
- #1079: web chat resolves the real author name from the cohort roster in
  mapMessages instead of always showing "Anonymous".

#1073, #1076, #1078 reviewed and closed as not-applicable.
