+++ ctf/docs/developer/ctf-plugin-feature-inventories/ctf-peer-programming-feature-inventory.md
@@ -197,6 +197,7 @@ Deterministic PeerProgramming seed script: `ctf/scripts/seedPeerProgramming.mjs`
 
 ## Change Log
 
+- 2026-06-27: **Code-review sweep follow-ups (#1074, #1075, #1077, #1079, #1080).** No route, schema, or contract change. (1) Removed the duplicate `_lib.ts`: the canonical copy at `app/api/peer-programming/_lib.ts` was dead code (no route imported it) while every route imported the divergent `lib/peer-programming/_lib.ts`, whose `peerProgrammingErrorResponse` only mapped `assignment_not_found` and fell through to 503 for everything else. The surviving `lib/` copy now maps all error codes (`invalid_payload` → 400, `policy_denied` → 403, `assignment_not_found`/`not_found` → 404, else `reportError` + 503), so member/admin routes return the right status instead of 503; the unused `app/api/` copy was deleted. (2) `POST /admin/single-open-cohort` now treats a missing `enabled` key the same as `null` (clear the setting), so an empty body `{}` no longer 400s — matches the contract's optional `enabled`. (3) Added server-side max-length checks (returning 400) on `POST /messages` and `POST /messages/[messageId]/replies` against `PEER_PROGRAMMING_MAX_MESSAGE_LENGTH` (2000), and on `POST /feedback` against `PEER_PROGRAMMING_MAX_FEEDBACK_LENGTH` (1000); these constants existed but were never enforced. (4) Web chat now shows the real author name: `mapMessages` resolves each message's author from the cohort roster (`@username`, or a short `Member <id>` fallback, matching the mobile helper) and sets `Message.author`, so the chat no longer labels every message "Anonymous". Sweep findings #1073, #1076, #1078 reviewed and closed as not-applicable (membership already verified by the `getMyCohort` join; the access gate already rejects empty `userId`; the mobile admin screen already shows a spinner until the admin check returns, with no client-side role available to gate on).
 - 2026-06-26: **Fixed the room failing to load ("Failed to load room") under single standing Cohort 1 mode.** The user-facing room page blanked to that error because `GET /api/peer-programming/room` returned non-OK and the shell hard-fails the whole page when the room fetch is non-OK. Two scale problems, both introduced/amplified by single standing Cohort 1 mode concentrating every active member and all their messages into one ever-growing cohort: (1) `peer_programming_messages` had no index on `cohort_id`, so `listMessages` (`WHERE cohort_id = $1 ORDER BY created_at ASC LIMIT 300`) did a sequential scan + sort over the whole table that eventually exceeded the DB statement timeout and threw → 503; (2) the room roster resolved **every** member's username via an external Clerk lookup on every load, unbounded work that also grew with the cohort. Fixes: added `idx_peer_programming_messages_cohort_created ON (cohort_id, created_at)` and `idx_peer_programming_cohort_members_cohort_created ON (cohort_id, created_at)` to `schema.sql` (`CREATE INDEX IF NOT EXISTS`, idempotent); capped the room roster to the earliest `PEER_PROGRAMMING_ROOM_ROSTER_LIMIT` (60) members (`buildCohortRosters` gained an optional `limitPerCohort`; the displayed total still comes from the true `cohort.memberCount`, so the count is unaffected — only the chip list is capped); and wrapped the roster build in the room route in try/catch so a slow or failing Clerk lookup degrades to an empty roster (logged via `reportError`) instead of failing the whole room — the same defense-in-depth applied to the Workforce dashboard. No route, contract, or response-shape change. Mirrors the earlier Workforce dashboard 503 fix.
 - 2026-06-26: **In-app admin toggle for single standing, always-open Cohort 1 mode.** The mode is no longer driven only by the env flag. Added a one-row settings singleton `peer_programming_settings` (`singleton_id BOOLEAN PRIMARY KEY DEFAULT TRUE CHECK (singleton_id)`, `single_open_cohort_enabled BOOLEAN` nullable, `updated_by_user_id TEXT`, `updated_at TIMESTAMPTZ`) to `schema.sql` and `schema.demo.sql` with the `CREATE TABLE IF NOT EXISTS` + `ALTER TABLE IF EXISTS … ADD COLUMN IF NOT EXISTS` pattern. New repository helpers `getPeerProgrammingSettings`, `setPeerProgrammingSingleOpenCohort`, and the async resolver `resolveSingleOpenCohortMode()` / `isSingleOpenCohortModeEnabled()` resolve the effective mode with precedence: (a) the persisted admin setting if set → use it; (b) else the env flag `PEER_PROGRAMMING_SINGLE_OPEN_COHORT` (read by the now env-only fallback `isPeerProgrammingSingleOpenCohortEnabled()` in `constants.ts`); (c) else default ON. The three behaviour call sites `getMyCohort`, `listActiveCohorts`, and `runWeeklyAssignment` now `await isSingleOpenCohortModeEnabled()` instead of the pure env read; behaviour with no admin setting and no env override is unchanged (ON). New admin route `app/api/peer-programming/admin/single-open-cohort/route.ts`: `GET` returns `{ enabled, source, adminSetting, envFlagEnabled }`; CSRF-guarded admin-only `POST { enabled: true|false|null }` upserts the singleton (null clears the admin setting) and audits as `peer-programming.settings.single-open-cohort.set`. Added a "Single standing Cohort 1 mode" control to `pp-admin-shell.tsx` (status badge, source label, Turn on / Turn off / Clear override, loading/error/success states). New command contracts `admin.single-open-cohort.get`/`.set`, matching access-policy and audit-event entries. Updated rule 123 to note the env flag is now a fallback under the admin setting. Web typecheck + eslint + EOF + inventory-drift checks clean.
 - 2026-06-26: **Single standing, always-open Cohort 1 mode (owner directive, low population).** Added the server flag `PEER_PROGRAMMING_SINGLE_OPEN_COHORT`, read by one resolver `isPeerProgrammingSingleOpenCohortEnabled()` in `lib/peer-programming/constants.ts`; it **defaults ON** (unset/empty = ON; only `0`/`false` turns it OFF). Added an `is_standing BOOLEAN NOT NULL DEFAULT FALSE` column to `peer_programming_cohorts` in `schema.sql` and `schema.demo.sql` (`ALTER TABLE IF EXISTS … ADD COLUMN IF NOT EXISTS`) plus a partial-unique index `uq_peer_programming_cohorts_standing ON (is_standing) WHERE is_standing` so there can only ever be one standing row. New repository helper `ensureStandingCohort(actorId)` find-or-creates the single standing cohort (label `C1`, `fallback_open = TRUE`, `is_standing = TRUE`), idempotent and race-safe via `ON CONFLICT (is_standing) WHERE is_standing DO NOTHING`. When the mode is ON: `getMyCohort` resolves the standing cohort (regardless of week) and idempotently joins the requesting (already access-gated) member so any active member can post, not just listen; `listActiveCohorts`/`getCohortById`/`listManagedCohorts` include the standing cohort regardless of week; and `runWeeklyAssignment` (cron + admin manual run) pauses the weekly split — it ensures the standing cohort exists and idempotently joins all provided active members into it with the same per-user+week notifications, creating no `C2`/`C3` (returns `cohortsCreated: 1`). When the flag is OFF the original weekly cohorting is byte-for-byte unchanged. Documented the flag in `.claude/rules/123-environment-configuration-rules.mdc`. Updated the seed script to set `is_standing` explicitly. No new API route or command added. Web typecheck + eslint + EOF + inventory-drift checks clean.
